refactor(DB): id 타입 INT에서 BIGINT로 변경

+++ src/main/resources/data.sql
@@ -1,19 +1,19 @@
 CREATE TABLE GAME (
-    id          INT         NOT NULL AUTO_INCREMENT,
+    id          BIGINT         NOT NULL AUTO_INCREMENT,
     trial_count INT         NOT NULL,
     created_at  DATETIME    NOT NULL default current_timestamp,
     PRIMARY KEY (id)
 );
 
 CREATE TABLE PLAYER (
-    id          INT         NOT NULL AUTO_INCREMENT,
+    id          BIGINT         NOT NULL AUTO_INCREMENT,
     name        varchar(10) NOT NULL,
     PRIMARY KEY (id)
 );
 
 CREATE TABLE PARTICIPATES (
-    game_id     INT     NOT NULL,
-    player_id   INT     NOT NULL,
+    game_id     BIGINT     NOT NULL,
+    player_id   BIGINT     NOT NULL,
     position    INT     NOT NULL,
     is_winner   BOOLEAN NOT NULL default false,
     PRIMARY KEY (game_id, player_id),
